makes getIsToken spec return false if spec is not parseable
