+++ src/route/index.ts
@@ -462,17 +462,9 @@ export async function initApiServer(
           }
 
           try {
-            const { result, error } = await getIsTokenSpec(
-              contractId,
-              network,
-              logger,
-            );
+            const isToken = await getIsTokenSpec(contractId, network, logger);
 
-            if (error) {
-              reply.code(400).send({ error, result: null });
-            } else {
-              reply.code(200).send({ data: result, error: null });
-            }
+            reply.code(200).send({ data: isToken, error: null });
           } catch (error) {
             reply.code(500).send("Unexpected Server Error");
           }
@@ -621,12 +613,10 @@ export async function initApiServer(
               return reply.code(500).send(ERROR.SERVER_ERROR);
             }
           }
-          return reply
-            .code(200)
-            .send({
-              data: { status: "miss" },
-              error: ERROR.SCAN_SITE_DISABLED,
-            });
+          return reply.code(200).send({
+            data: { status: "miss" },
+            error: ERROR.SCAN_SITE_DISABLED,
+          });
         },
       });
